Move namespaces to root tag when adding an attribute.
When the layout editor is adding a non android attribute it will
ensure the app namespace is declared on the root tag in the layout
file.

With the addition of the <layout> tag for data binding people often
have added an empty <layout> tag around an existing layout file
such that the namespace declarations appear on a different tag than
the root tag (the layout).

This CL will move the namespaces out of the tag that is being
changed to the root tag.

Also: fix a bug in the code for moving the namespace. If the root
is updated with an attribute value the existing code would remove
all but the newly added namespace.

Fixes: 120507391
Test: Unit tests added.

+++ designer/src/com/android/tools/idea/common/model/NlComponent.java
@@ -399,6 +399,7 @@ public class NlComponent implements NlAttributesHolder {
 
     String prefix = null;
     if (namespace != null && !ANDROID_URI.equals(namespace)) {
+      transferNamespaces(this);
       prefix = AndroidResourceUtil.ensureNamespaceImported((XmlFile)tag.getContainingFile(), namespace, null);
     }
     String previous = getAttribute(namespace, attribute);
@@ -519,6 +520,7 @@ public class NlComponent implements NlAttributesHolder {
     if (tag == null) {
       return null;
     }
+    transferNamespaces(this);
     return AndroidResourceUtil.ensureNamespaceImported((XmlFile)tag.getContainingFile(), namespace, suggestedPrefix);
   }
 
@@ -799,17 +801,20 @@ public class NlComponent implements NlAttributesHolder {
   }
 
   /**
-   * Given a tag on the current component which is not yet part of the current
-   * document and a receiver component where the tag is going to be added:
+   * Transfer the namespace declarations to the root of the document.
+   *
+   * The current component may not be part of the current document yet
+   * (if this happens while adding components) and a receiver component
+   * where the tag is going to be added:
    * <ul>
    *   <li>look up any namespaces defined on the receiver or its parents</li>
    *   <li>look up any namespaces defined on the current new tag</li>
    * </ul>
-   * and transfer all those namespace declarations to the current document
+   * and transfer all those namespace declarations to the current document root.
    */
   private void transferNamespaces(@NotNull NlComponent receiver) {
     XmlTag rootTag = getDocumentRoot();
-    XmlTag tag = receiver.getTagDeprecated();
+    XmlTag tag = receiver.getTag();
     while (tag != null && tag != rootTag) {
       if (!tag.getLocalNamespaceDeclarations().isEmpty()) {
         // This is done to cleanup after a manual change of the Xml file.
@@ -818,7 +823,7 @@ public class NlComponent implements NlAttributesHolder {
       }
 
       receiver = receiver.getParent();
-      tag = receiver != null ? receiver.getTagDeprecated() : null;
+      tag = receiver != null ? receiver.getTag() : null;
     }
     transferLocalNamespaces();
   }
@@ -832,7 +837,8 @@ public class NlComponent implements NlAttributesHolder {
    */
   private void transferLocalNamespaces() {
     XmlTag rootTag = getDocumentRoot();
-    if (rootTag == this) {
+    XmlTag tag = getTag();
+    if (tag == null || rootTag == tag) {
       return;
     }
     // Transfer namespace attributes to the root tag
@@ -843,7 +849,7 @@ public class NlComponent implements NlAttributesHolder {
     }
     Map<String, String> oldPrefixToPrefix = Maps.newHashMap();
 
-    for (Map.Entry<String, String> entry : getTagDeprecated().getLocalNamespaceDeclarations().entrySet()) {
+    for (Map.Entry<String, String> entry : tag.getLocalNamespaceDeclarations().entrySet()) {
       String namespace = entry.getValue();
       String prefix = entry.getKey();
       String currentPrefix = namespaceToPrefix.get(namespace);
@@ -866,7 +872,8 @@ public class NlComponent implements NlAttributesHolder {
     }
 
     if (!oldPrefixToPrefix.isEmpty()) {
-      updatePrefixes(getTagDeprecated(), oldPrefixToPrefix);
+
+      updatePrefixes(tag, oldPrefixToPrefix);
     }
 
     removeNamespaceAttributes();
@@ -911,9 +918,12 @@ public class NlComponent implements NlAttributesHolder {
   }
 
   private void removeNamespaceAttributes() {
-    for (XmlAttribute attribute : getTagDeprecated().getAttributes()) {
-      if (attribute.getName().startsWith(XMLNS_PREFIX)) {
-        attribute.delete();
+    XmlTag tag = getTag();
+    if (tag != null) {
+      for (XmlAttribute attribute : tag.getAttributes()) {
+        if (attribute.getName().startsWith(XMLNS_PREFIX)) {
+          attribute.delete();
+        }
       }
     }
   }
